requirements.txt
dt_apriltags==3.1.7
Jetson.GPIO==2.1.6
matplotlib==3.1.2
numpy==1.24.4
openai==1.54.4
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==11.0.0
PyAudio==0.2.14
pymycobot==3.4.9
PyYAML==6.0.2
Requests==2.32.3
scipy==1.3.3
soundfile==0.12.1
tf==1.0.3
transforms3d==0.4.2
